+++ test/integration/test_cases/web_content/cookbook/rendering/grid_lines.mjs
@@ -9,8 +9,9 @@ const testSteps = [
 	chart => {
 
 		let toCanvasRect = (rect) => {
-			let pos = chart._toCanvasCoords({ x: rect.pos.x, y: rect.pos.y + rect.size.y });
-			let pos2 = chart._toCanvasCoords({ x: rect.pos.x + rect.size.x, y: rect.pos.y });
+			let convert = chart.getConverter("plot-area", "relative", "canvas");
+			let pos = convert({ x: rect.pos.x, y: rect.pos.y + rect.size.y });
+			let pos2 = convert({ x: rect.pos.x + rect.size.x, y: rect.pos.y });
 			return { pos, size: { x: pos2.x - pos.x, y: pos2.y - pos.y } };
 		}
 
@@ -25,7 +26,9 @@ const testSteps = [
 
 			ctx.lineWidth = 1;
 
-			if (chart._toRelCoords({ x: 0, y: rect.pos.y }).y !== 1)
+			let convert = chart.getConverter("plot-area", "canvas", "relative");
+
+			if (convert({ x: 0, y: rect.pos.y }).y !== 1)
 			{
 				ctx.beginPath();
 				ctx.moveTo(rect.pos.x, rect.pos.y);
@@ -33,7 +36,7 @@ const testSteps = [
 				ctx.stroke();	
 			}
 
-			if (chart._toRelCoords({ x: 0, y: rect.pos.y + rect.size.y }).y !== 1)
+			if (convert({ x: 0, y: rect.pos.y + rect.size.y }).y !== 1)
 			{
 				ctx.beginPath();
 				ctx.moveTo(rect.pos.x, rect.pos.y + rect.size.y);